Flatpickrのテストを共通化する
